Proof-of-concept plug-in for files on Acacia object store
This small proof-of-concept set of modules shows how a pair of plug-ins
can be used to copy all incoming files into a given bucket of the Acacia
object store, while also serving all retrieve requests by first staging
the file back from Acacia into the local filesystem.

To run one needs to export the RCLONE_ACACIA_REMOTE and ACACIA_BUCKET
environment variables to point to the correct rclone remote definition,
and the bucket where data will be synchronised to/from. These
environment variables are needed mostly because staging plug-ins can't
currently be given a parameter set when they are created, so these
settings cannot be put into the NGAS configuration file. With some small
amount of work this could be improved on the NGAS side.

NGAS required some small (backwards-compatible) modifications to make
this work. Make sure you are using the latest (as the time of writing)
NGAS server.

This is a fairly simplistic solution, and highly performance-inadequate.
Some points to consider before declaring any of this production ready:

 * The local files are never deleted, even after copying them to Acacia.
 * Copying to acacia is synchronous so it slows down the *ARCHIVE
   request. It should be asynchronous, with a queue, status check, etc.
 * After copying files to Acacia we should remove them locally. One way
   is correctly configuring the cache control thread, but I don't
   remember now all the details required to do that sensibly. In
   particular, we still want the files to be listed in the DB. Another
   option is directly removing the file from disk using `os.unlink`, but
   again I don't remember if that triggers some other corrective actions
   in the DB in the background.
 * Staging is synchronous, but that's how it's always been.
 * This is using rclone, we could use boto3 eventually.
 * File staging and archiving *always* goes through the local filesystem
   first. Doing a direct streaming into/from Acacia would be *very*
   difficult because the NGAS deep in its design always assumes a
   filesystem.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>